# openapi.yml
openapi: "3.0.3"
info:
  title: "sample api"
  version: "1.0.0"
  description: "ユーザのCRUDのAPI"
servers:
  - url: 'http://localhost:8080/'
    description: 'ローカル環境'
paths:
  "/users":
    get:
      summary: "全ユーザ取得"
      tags:
        - User
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
    post:
      summary: "ユーザ作成"
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              example: 
                name: "次郎"
      responses:
        "200":
          description: "成功"
  "/users/{id}":
    get:
      summary: "ユーザ取得"
      tags:
        - User
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
          example: 10
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: "ユーザが見つからない"
    put:
      summary: "ユーザ更新"
      tags:
        - User
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
          example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              example:
                name: "二郎"
      responses:
        "200":
          description: "成功"
        "404":
          description: "ユーザが見つからない"
    delete:
      summary: "ユーザ削除"
      tags:
        - User
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
          example: 10
      responses:
        "200":
          description: "成功"
        "404":
          description: "ユーザが見つからない"
components:
  schemas:
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 10
        name:
          type: "string"
          example: "太郎"
    Users:
      type: "array"
      items:
        type: "object"
        properties:
          id:
            type: "integer"
            format: "int64"
            example: 10
          name:
            type: "string"
            example: "太郎"